Update EWAH from upstream lemire/EWAHBoolArray
