+++ ewah_bool_utils/__init__.py
@@ -1,5 +1,4 @@
 """Top-level package for EWAH Bool Utils."""
 
-__version__ = '0.1.0'
 
 from .ewah_bool_wrap import *
